@@ -1,10 +1,10 @@
 name: currency_text_input_formatter
 description: Currency Text Input Formatter for Flutter. Use it easy and simple for your flutter app.
-version: 1.0.5
+version: 2.0.0
 homepage: https://github.com/gtgalone/currency_text_input_formatter
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
